refactor model init into submodules
